+++ crates/ruma-events/src/lib.rs
@@ -162,6 +162,8 @@ pub mod key;
 #[cfg(feature = "unstable-msc3488")]
 pub mod location;
 pub mod marked_unread;
+#[cfg(feature = "unstable-msc4171")]
+pub mod member_hints;
 #[cfg(feature = "unstable-msc1767")]
 pub mod message;
 #[cfg(feature = "unstable-pdu")]